Finished the move to store the FeatureType information in the respository, and the use of NodeRefs contaning full paths instead of lists of single steps.
So far there were no FeatureType information in the repo as FeatureTypes were provided by the geoserver plugin from the original prototype.
Now tree NodeRef's have a metadataId property, which points to the ObjectId of the stored RevFeatureType.
RevFeatureType is a new kind of RevObject, as well as RevFeature. Using them allows to decouple the core from GeoTools, abstracting it out
of the feature model in use. GeoToolsRevFeature  and GeoToolsRevFeatureType exist in org.geogit.storage.hessian and in core are only used by unit/integration tests.
This patch also introduces the usage of com.google.common.base.Optional<T> as the return type of various query methods that could return no-data,
in order to make the API more appealing and less prone to programming errors by getting rid of null as a valid return type. That work shall continue as the API gets revised.