Refactor ObjectDatabase to deal with ObjectSerialisingFactory internally.
This commit is a refactoring of ObjectDatabase so that we don't need to pass ObjectSerialisingFactory
all the way around but instead makes the ODB API much friendlier. The IoC engine should inject ObjectSerialisingFactory
into the ObjectDatabase implementations as needed, and the API ends up being as simple as put(RevObject) instead of
put(ObjectId, ObjectSerialisingFactory.createXXXWriter(revObject)).