Initial implementation of cherry-pick.
This patch adds the porcelain CherryPickOp to core and the cli cherry-pick command.
Note however that it'll apply the changes from the supplied commit as a whole, meaning
that if there are non conflicting edits to a given feature at both ends then the current
state will be overritten. This is because so far we can only diff as whole objects (see
DiffEntry). In the future we'll need to handle conflicts and non conflicting merges properly.