Support commit message and author hooks in GeoTools transaction object
GeoServer has a 'convention' of setting the commit message and name of the user issuing a transaction as properties in
the GeoTools Transaction object, named VersioningCommitMessage and VersioningCommitAuthor respectively. This patch
adds support to grabing that information and put it on the Geogit commit when a transaction is committed
at GeogitTransactionState.commit().