Encapsulate AbstractGeogitCommand's workingTree, refDatabase, and index.
This patch makes commands access the working tree, index, and ref database consistently
through CommandLocator and encapsulates the behavior in three protected getters
on AbstractCommandLocator. This way it doesn't matter (for commands and other code)
whether the CommandLocator comes from and hence no need to figure out if its a transaction
or not to get the working tree/index/refdb from different places.