Some more simple code/style changes after merging branch jdgarrett/CommonAncestorFix
- Add missing class headers
- ObjectDatabasePutInterceptor intercepts putAll with a lighter weight approach
at figuring out which RevObjects in the argument iterator are RevCommits. It used
to defeat the purpose of the argument being an iterator by building an List out of
it, now uses a decorating iterator that adds only commits to the list to be added
to the graph db once the actual putAll call succeeded.
- Rename GraphDatabase.lowestCommonAncestor(...) as findLowestCommonAncestor(...)