Declare RevCommit.getParentIds() to return ImmutableList instead of just List.
RevCommit is supposed to be immutable so it doesn't make sense to have its list
of parent ids as a mutable list and having to make defensive copies when returning
it, nor when client code uses the returned list. Declaring the return type to
be an ImmutableList also serves as documentation that the list is immutable and
that it doesn't accept/contains nulls.