Boolean literals should never be used in equality tests.
Comparing the result of an expression to the literal value "false"
is not necessary and usually makes the code harder to read. Consider
negating the expression and removing the equality test.

Comparing the result of an expression to the literal value "true"
is not necessary and usually makes the code harder to read. Consider
removing the equality test.