Updated the web-api to support long transactions and added the Merge web api command.
Changed all web-api commands to support running off of a transaction.
Added Merge which supports checking for conflicts, a dry-run option which sends back what will change in the merge and merge itself with the no-commit option.
Fixed a bug in ReportMergeScenarioOp where removed features weren't being added to the unconflicted list.
The previously stated bug fix caused a test to fail which revealed a problem with the RevTreeImpl class where if the tree was empty when it was read in it would be created with an empty bucket list instead of Optional.absent() which was causing it to list a feature as removed twice when DiffTree was called.