Run geogit with the GeoTools force XY axis order hint on.
Check out the following link for an explanation of what this means:
http://docs.geotools.org/latest/userguide/library/referencing/order.html

This fixes the problem of importing PostGIS layers in a geographic CRS,
which resulted in inverted axis orders. Yet, the global hint strikes
me as a bad solution, given that for this specific case it should be
the GeoTools PostGIS DataStore who forces the XY axis order in its
FeatureTypes, given PostGIS is known to define them in that order.