Updated CheckoutWebOp and PullWebOp to support conflict resolution.
Updated CheckoutWebOp to support ours and theirs options to resolve conflicts with.
Updated PullWebOp to return the merge commit result so that we can show conflicts if they should arise.
Updated MergeReport to store the objectIds of ours, theirs and ancestor commits.
Fixed a problem with TransactionEnd so that it didn't stop if there was a NothingToCommitException in the case of a fast-forward.