"Fix an issue that caused FeatureStore edits to change the metadataId of RevFeature nodes.
Geogit supports adding features of disparate feature types as direct children of the same tree.
This shadowed an issue with the geogit FeatureStore implementation that was causing to store
feature nodes with their own metadataId instead of blank (and hence defer to the tree's default one)
when a namespace is assigned to the DataStore.
This patch makes sure both inserts and updates through the geotools api use a feature with the
native/internal FeatureType and hence ensures their metadataId will remain loyal to the tree's default.