OSM download bug fixes.
Fixed a bug where update would always base the new commit off of the original download commit
causing the same changes to show up every time.
Fixed a bug where the user would be left on OSM_FETCH if there were no changes in the update.
Fixed a bug where the update would not be merged into the current branch and the user would be
left on OSM_FETCH.