Use SLF4J with Logback backend as the logging framework.
The core module depends only on the slf4j API.
The cli app adds logback as a dependency specific to cli
and logs to the .geogit/log/geogit.log file.
The logback configuration file is loaded from .geogit/log/logback.xml
and a default one is created if it doesn't exist.
The location of the log file is specified in logback.xml
using the ${geogitdir} variable whose value is set to the logback
context with the result of the resolve-geogit-dir command. This way
the log file location doesn't need to be an absolute path in logback.xml
and can resolve to the correct location regardless of being at the
repository's top level directory or not.
The org.slf4j.jul-to-slf4j library is used to redirect java.util.logging
loggers (e.g. used by geotools) to logback and hence to geogit.log