Mark nodes to be deleted as having NULL ObjectId.
Instead of having RevTreeBuilder2 an in memory list of nodes to be deleted,
make it mark them as having ObjectId == NULL and add them to the NodeIndex,
then call RevTreeBuilder.remove(node.getName()) instead of RevTreeBuilder.put(),
to avoid potential out of memory errors.

Also make FeatureToDelete hold its type instead of a separate "path" field
to make it less of a special case.