Fix Ref.getType() to return COMMIT, deprecate it, and get rid of its uses everywhere.
Ref.getType() is kind of a forced property. It doesn't really belong to the
ref database and its practically hardcoded to be COMMIT.
The Ref.getType() method gets deprecated and scheduled to be removed in 0.9.0
as its practically of no use and forces ref-parse to unnecessarily call on
resolve-object-type, which can be called separately in case some client code
needs to figure out what kind of object a ref points to with the ref's object
id as argument.