Remove duplication of code between geogit-geoserver and web-app
All web commands were moved to web-api to be used by noth web-app
and geogit-geoserver, each of which provide an implementation of
the new RepositoryProvider interface for web commands to get the
geogit instance to work with.
geogit-geoserver uses a RepositoryProvider that grabs geogit instances
from geogit datastores in the geoserver catalog.
web-app uses a simple RepositoryProvider for the single geogit
instance its serving.