Fix missing commits bug in the default graph database.
Makes TinkerGraphDatabase.Impl implement BlueprintsGraphDatabase.commit()
so that the graph is saved to disk on each update. Otherwise all the
changes were only in memory until TinkerGraphDatabase.close() was
properly called, which frequently leads to missing entries in the graph
database.

A test case was added to TinkerGraphDatabaseTest, and the abstract
test suites for graph databases were improved so that they don't require
a full repository to be run.